RCAagent/specializedAgents/ContextualScout/mcp/lambda/requirements.txt
# Add requirements needed for the lambda handler. For example boto3
tavily-python>=0.5.0